Fix integer overflow in Aversive's path finder
It uses uin8_t, because that made sense on ATMegas back in the days.
However it causes issues as soon as we have a few rays cast. Starting
from even 4 squares obstacles could be enough to create issues. Now that
we have integers, it should be better!

+++ master-firmware/src/aversive/math/geometry/circles.c
@@ -57,7 +57,7 @@ int circle_intersect(const circle_t *c1, const circle_t *c2,
 {
     circle_t ca, cb;
     float a, b, c, d, e;
-    uint8_t ret = 0;
+    int ret = 0;
 
 
     /* We have to assume that either delta_x or delta_y is not zero to avoid
